+++ app/src/test/java/org/se13/utils/MatrixTest.java
@@ -7,7 +7,7 @@ class MatrixTest {
 
     @Test
     void randnTest() {
-        double[][] matrix = Matrix.randn(3, 2);
+        float[][] matrix = Matrix.randn(3, 2);
 
         Assertions.assertEquals(3, matrix.length);
         Assertions.assertEquals(2, matrix[0].length);
@@ -21,20 +21,20 @@ class MatrixTest {
 
     @Test
     void matmulTest() {
-        double[][] a = new double[][]{
-            new double[]{1, 2},
-            new double[]{3, 4},
+        float[][] a = new float[][]{
+            new float[]{1, 2},
+            new float[]{3, 4},
         };
-        double[][] b = new double[][]{
-            new double[]{5, 6, 7},
-            new double[]{8, 9, 10},
+        float[][] b = new float[][]{
+            new float[]{5, 6, 7},
+            new float[]{8, 9, 10},
         };
-        double[][] c = new double[][]{
-            new double[]{21, 24, 27},
-            new double[]{47, 54, 61},
+        float[][] c = new float[][]{
+            new float[]{21, 24, 27},
+            new float[]{47, 54, 61},
         };
 
-        double[][] result = Matrix.matmul(a, b);
+        float[][] result = Matrix.matmul(a, b);
 
         for (int i = 0; i < c.length; i++) {
             for (int j = 0; j < c[i].length; j++) {
@@ -45,22 +45,22 @@ class MatrixTest {
 
     @Test
     void reluTest() {
-        double[][] a = new double[][]{
-            new double[]{-0.5, 0.5},
-            new double[]{0.5, -0.5},
+        float[][] a = new float[][]{
+            new float[]{-0.5f, 0.5f},
+            new float[]{0.5f, -0.5f},
         };
 
-        double[][] result = Matrix.relu(a);
+        float[][] result = Matrix.relu(a);
 
-        Assertions.assertEquals(result[0][0], 0);
-        Assertions.assertEquals(result[0][1], 0.5);
-        Assertions.assertEquals(result[1][0], 0.5);
-        Assertions.assertEquals(result[1][1], 0);
+        Assertions.assertEquals(result[0][0], 0f);
+        Assertions.assertEquals(result[0][1], 0.5f);
+        Assertions.assertEquals(result[1][0], 0.5f);
+        Assertions.assertEquals(result[1][1], 0f);
     }
 
     @Test
     void argmaxTest() {
-        double[] a = new double[]{1, 3, 7, 4};
+        float[] a = new float[]{1, 3, 7, 4};
         Assertions.assertEquals(2, Matrix.argmax(a));
     }
 }
